refactor: delete AI News module entirely
Remove the AI HOT third-party aggregation module: API client, store,
models, cache, refresh policy (renamed PopoverInteractionPolicy keeps
hover/layout policies), all enum cases, prefs key, UI surfaces, tests,
smoke seeds, and doc mentions. Stored enabledModules/primaryFavorites
entries self-normalize; cache file left orphaned.
