+++ README.md
@@ -52,7 +52,6 @@ No `.app.zip` on Releases (unsigned browser downloads trip Gatekeeper). Or run `
 | **Work / Break** | off | Focus timer, menu-bar countdown, break overlay |
 | **System** | off | CPU / memory, disk categories, top processes |
 | **Goals** | off | Today / Week / Vision / Schedule, notes, tags, heatmap |
-| **AI News** | off | AI HOT Top 10, hover summaries, source context |
 
 Theme: **Mono** only (black / white / gray, light & dark).
