+++ Sources/Kaji/Prefs.swift
@@ -77,13 +77,6 @@ final class Prefs: ObservableObject {
     @Published var preventSleep: Bool {
         didSet { defaults.set(preventSleep, forKey: Key.preventSleep) }
     }
-    @Published var aiNewsRefreshHours: Int {
-        didSet {
-            let normalized = AIHotRefreshPolicy.normalize(hours: aiNewsRefreshHours)
-            if normalized != aiNewsRefreshHours { aiNewsRefreshHours = normalized; return }
-            defaults.set(normalized, forKey: Key.aiNewsRefreshHours)
-        }
-    }
     @Published var mailBriefHour: Int {
         didSet { defaults.set(min(23, max(0, mailBriefHour)), forKey: Key.mailBriefHour) }
     }
@@ -127,7 +120,6 @@ final class Prefs: ObservableObject {
         static let launchAtLogin = "launchAtLogin"
         static let preventSleep = "preventSleep"
         static let preferencesInitialized = "preferencesInitialized"
-        static let aiNewsRefreshHours = "aiNewsRefreshHours"
         static let mailBriefHour = "mailBriefHour"
         static let mailBriefMinute = "mailBriefMinute"
         static let mailBriefBatchSize = "mailBriefBatchSize"
@@ -140,7 +132,7 @@ final class Prefs: ObservableObject {
             showRemaining, focusMinutes, breakMinutes,
             allowBreakSkip, breakOverlayEnabled, visibleProvidersV2,
             visibleProvidersCursor, autoCleanEnabled, launchAtLogin, preventSleep,
-            aiNewsRefreshHours, mailBriefHour, mailBriefMinute,
+            mailBriefHour, mailBriefMinute,
             mailBriefBatchSize, mailBriefConcurrency, mailBriefModel, goalGrouping,
             primaryFavorites, showsWorkCompactSignal
         ]
@@ -248,14 +240,12 @@ final class Prefs: ObservableObject {
         } else {
             preventSleep = false
         }
-        aiNewsRefreshHours = AIHotRefreshPolicy.normalize(hours: d.object(forKey: Key.aiNewsRefreshHours) == nil ? 5 : d.integer(forKey: Key.aiNewsRefreshHours))
         mailBriefHour = d.object(forKey: Key.mailBriefHour) == nil ? 9 : min(23, max(0, d.integer(forKey: Key.mailBriefHour)))
         mailBriefMinute = d.object(forKey: Key.mailBriefMinute) == nil ? 0 : min(59, max(0, d.integer(forKey: Key.mailBriefMinute)))
         mailBriefBatchSize = MailBriefBatchPolicy.batchSize(d.object(forKey: Key.mailBriefBatchSize) == nil ? 10 : d.integer(forKey: Key.mailBriefBatchSize))
         mailBriefConcurrency = MailBriefBatchPolicy.concurrency(d.object(forKey: Key.mailBriefConcurrency) == nil ? 2 : d.integer(forKey: Key.mailBriefConcurrency))
         mailBriefModel = MailBriefModel.normalize(d.string(forKey: Key.mailBriefModel))
         d.set(mailBriefModel.rawValue, forKey: Key.mailBriefModel)
-        d.set(aiNewsRefreshHours, forKey: Key.aiNewsRefreshHours)
         d.set(true, forKey: Key.preferencesInitialized)
     }
