+++ Sources/Kaji/SettingsView.swift
@@ -8,7 +8,6 @@ enum SettingsSection: String, CaseIterable, Identifiable {
     case work = "Work"
     case goals = "Goals"
     case quota = "Quota"
-    case aiNews = "AI News"
     case mailBrief = "Mail Brief"
     case cli = "CLI"
     case permissions = "Permissions"
@@ -21,7 +20,6 @@ enum SettingsSection: String, CaseIterable, Identifiable {
         case .work: "timer"
         case .goals: "checkmark.circle"
         case .quota: "gauge.with.dots.needle.67percent"
-        case .aiNews: "newspaper"
         case .mailBrief: "envelope"
         case .cli: "terminal"
         case .permissions: "lock.shield"
@@ -87,7 +85,6 @@ struct SettingsView: View {
                 return true
             case .work: return prefs.isModuleEnabled(.work)
             case .goals: return prefs.isModuleEnabled(.goals)
-            case .aiNews: return prefs.isModuleEnabled(.aiNews)
             case .mailBrief: return prefs.isModuleEnabled(.mailBrief)
             }
         }
@@ -165,7 +162,6 @@ struct SettingsView: View {
                     moduleRow(.work, title: L10n.t(.moduleWork, prefs.language), lockedOn: false)
                     moduleRow(.system, title: L10n.t(.moduleSystem, prefs.language), lockedOn: false)
                     moduleRow(.goals, title: L10n.t(.moduleGoals, prefs.language), lockedOn: false)
-                    moduleRow(.aiNews, title: L10n.t(.moduleAINews, prefs.language), lockedOn: false)
                     moduleRow(.mailBrief, title: "Mail Brief", lockedOn: false)
                     moduleRow(.launchd, title: "Background Tasks", lockedOn: false)
                 }
@@ -325,20 +321,6 @@ struct SettingsView: View {
                         providerRow(key)
                     }
                 }
-            }
-            if selection == .aiNews {
-                settingBlock(title: "AI News") {
-                VStack(alignment: .leading, spacing: 7) {
-                    settingRow(title: L10n.t(.refreshInterval, prefs.language)) {
-                        ForEach(AIHotRefreshPolicy.allowedHours, id: \.self) { hours in
-                            segment("\(hours)h", on: prefs.aiNewsRefreshHours == hours) {
-                                prefs.aiNewsRefreshHours = hours
-                            }
-                        }
-                    }
-                    .help("AI News uses AI HOT's anonymous read-only v1 API.")
-                }
-            }
             }
             if selection == .mailBrief {
                 settingBlock(title: "Mail Brief") {
