+++ Sources/KajiCore/ModulePrefsLogic.swift
@@ -6,12 +6,11 @@ public enum KajiModuleID: String, CaseIterable, Codable, Sendable, Comparable {
     case work
     case system
     case goals
-    case aiNews
     case mailBrief
     case launchd
 
     /// Stable popover order. Unknown / disabled ids never appear here.
-    public static let stableOrder: [KajiModuleID] = [.quota, .work, .system, .goals, .aiNews, .mailBrief, .launchd]
+    public static let stableOrder: [KajiModuleID] = [.quota, .work, .system, .goals, .mailBrief, .launchd]
 
     public static func < (lhs: KajiModuleID, rhs: KajiModuleID) -> Bool {
         let li = stableOrder.firstIndex(of: lhs) ?? Int.max
